+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This version supports OneView appliances with version 4.2 using the OneView REST
 - Interconnect
 - Interconnect Type
 - Logical enclosure
+- Logical interconnect group
 - Network set
 - Scope
 - Server hardware